requirements.txt
osmnx
geopandas
pandas


aws-cdk-lib==2.110.0
constructs==10.3.0
cdk-ecr-deployment==3.0.15

python-dotenv==1.0.0

boto3==1.20.29
botocore==1.23.29
coverage

cdk_nag>=2.0.0